[ntp-xdtr] Connects handler with page content annotations
Retrieves the page content annotations for a given visit to determine
whether to show or hide a visit based on the visibility score in the NTP
XDTR module. Updates max number of tabs to show in module. Additionally, adds a flag to control the visibility score threshold. Also updates unit tests. Refactors test support so history clusters and tab resumption can share the same mock history service which is needed now that we are using page content annotations.

Note: Will add feature flag to control whether -1 is included or not
in the shown results in a follow up.

Bug: 1491267, 1512906
Change-Id: Iaf0bda57df19197b05d540473547c5688697cbec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5084000
Commit-Queue: Marlon Facey <mfacey@chromium.org>
Commit-Queue: Roman Arora <romanarora@chromium.org>
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Reviewed-by: Roman Arora <romanarora@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254196}