CrOS SampleSystemWebApp: Leverage build_webui(), use Mojo TS bindings.
This is a partial revive of http://crrev.com/c/4312587, which had been
heavily discussed about a year ago.

Firstly, SampleSystemWebApp is supposed to be used as an example for
CrOS WebUI SWA authors on how to structure their app, but it has ended
up being very different (in an obsolete way) to what is used in practice
under the ash/webui/ folder, where build_webui() has become the
standard, as well as Mojo TS bindings (as opposed to JS bindings).

Secondly, the SWA team that had concerns about this approach has been
increasingly less involved with this part of the code, and holding
back updating SampleSystemWebApp to a simpler (see negative  and more
standard form does not seem like a good strategy anymore.

At this point, it is not heplful to keep SampleSystemWebApp as-is,
and should either be fully deleted or updated to be a representative
example (as done by this CL).

Bug: 1002798
Change-Id: Ie8a09f2d08c965cccf9275768caf1b2568c6f85e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5246131
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254199}