Fix TestViewsExamplesLaunches when Fluent scrollbars are enabled
When Fluent scrollbars are enabled, the call to NativeTheme instantiates
NativeThemeFluent, which calls skia::DefaultFontMgr() in its
constructor.
Later, gfx::InitializeFonts() gets called in ExamplesMainProc, which
expects the font manager to not be initialized and crashes.
This CL initializes fonts when Fluent scrollbars are enabled before the
NativeTheme gets instantiated to avoid crashing the test.

Bug: 1502537
Change-Id: I6be639226ba39475f3254dc42d1db8dc60681256
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5111784
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Yaroslav Shalivskyy <yshalivskyy@microsoft.com>
Reviewed-by: Nasser Al-shawwa <alshawwa@chromium.org>
Commit-Queue: Gaston Rodriguez <gastonr@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1254210}