Remove xuhong@chromium.org from tools/metrics/histograms/metadata/platform/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/xuhong@chromium.org

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I9ea94933866b05f7be767ffba0a588d60b3a6638
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5250830
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254229}