[ Extensions ] Fix tabGroups.Move bug with adjacent groups
Fixes a bug that prevents extensions from moving a tab group to the right of another one due to an off by one error.

Fixing this caused a few other call sites to fail since they had implemented similar work-arounds.

The current way we calculate the index to move a tab group is by pretending the group and all of its tabs have been removed from the Tabstrip. The left over tabs represent the available indices we could move the group to (except for indices in the middle of another group).

Because there was an off by one error in the tab group extensions API and the case (moving a group to the right of another group) was unhandled in the TabstripModel, this has gone mostly unnoticed since their effects canceled each other out in most cases.

Regression tests are added in tab_groups_api_unittest for this scenario.

Change-Id: I92d8350157ab2ff23a6970bb5fe0ce4c15f5e881
Bug: 1417493
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5225325
Reviewed-by: Yuheng Huang <yuhengh@chromium.org>
Commit-Queue: Darryl James <dljames@chromium.org>
Reviewed-by: Taylor Bergquist <tbergquist@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254231}