Remove angusmclean@google.com from ash/system/palette/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/angusmclean@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I3823a2bb5f158cc6cf1cc724f3ba95a4e1c8d1eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5249814
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Reviewed-by: David Black <dmblack@google.com>
Commit-Queue: David Black <dmblack@google.com>
Cr-Commit-Position: refs/heads/main@{#1254237}