media/gpu/ImageProcessorPerfTest JSON test results
Previous to this CL, the ImageProcessorPerf test results were written to
std::out. The test data was thus not able to be read via tools such as
Crosbolt.

Updates the ImageProcessorPerf test to have an option to specify an
output location to write the corresponding logs. If not specified, it
will output to the current directory.

Each ImageProcessorPerf test will output a json file corresponding to
the measured metrics of each test. These json files will be stored in
the specified output directory.

Bug: b:316160886
Test: Ran ImageProcessorPerfTest on Trogdor.
Change-Id: I647956b27a80fd4e4132b9f74c8de4710e8cda8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5234559
Reviewed-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Auto-Submit: Bijan Choobineh‎ <bchoobineh@chromium.org>
Commit-Queue: Bijan Choobineh‎ <bchoobineh@chromium.org>
Reviewed-by: Justin Green <greenjustin@google.com>
Cr-Commit-Position: refs/heads/main@{#1254246}