Implement HeatmapPalmRejectionClient on the chromium side
Since we are now reading heatmap data in ML service, we no longer need to send palm detection request from chromium side. Instead, we need to make a load service request when we initialize the device. As a result, we need to rework the APIs in HeatmapPalmDetector.

In this CL, we implemented necessary API in HeatmapPalmDetector to make it a HeatmapPalmRejectionClient. We will remove the deprecated old API, such as DetectPalm, in a follow up CL.

Bug: b:315720763
Change-Id: Ic9f506c8c7d39189644fa7f5cd6011ee77ae495c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5248485
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Rob Schonberger <robsc@chromium.org>
Reviewed-by: Andrew Moylan <amoylan@chromium.org>
Reviewed-by: Darren Shen <shend@chromium.org>
Commit-Queue: Jing Wang <jiwan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254277}