[infra] Add edges to mixins wherever they are used.
Currently mixins specified in the config for a test within a basic suite
or on a test target itself do not have an edge created and are instead
just stored on a structure and copied into test_suites.pyl. This is in
contrast to their use in the matrix config in a matrix compound suite.
This CL changes it so that edges are created wherever a mixin is used
and those edges are used for generating the mixins and remove_mixins
fields. This prevents the currently possible situation where unused
tests can refer to non-existent mixins. Instead, all references to
mixins will be checked to exist when the starlark is run.

Bug: 1420012
Change-Id: Idf916fc7a8471b46e06462d50468bd27b4891737
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5249760
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1254280}