webview: try even harder to work around the resource problem.
The previously-landed workaround from
https://chromium-review.googlesource.com/c/chromium/src/+/3339195
was only used on S+ devices and only if WebView initialization was
triggered from a background thread, but since this was implemented it's
become clear that this is not the only situation where the workaround
might help, and the risk of trying it is extremely low as we already
test if the resources *are* missing before trying to re-add them - so
if we don't try the workaround, we are almost guaranteed to crash
anyway, just a few lines later when we first need to access a resource.

So, just try the existing workaround on all OS versions, regardless of
which thread WebView init was launched from. This definitely won't solve
all possible cases here, but seems likely to help with some.

Bug: 1272888
Change-Id: Iea085903666c0f52acfa5fc0e92a4417271600c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5231451
Auto-Submit: Richard (Torne) Coles <torne@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254281}