emoji_picker: Use DOM attributes instead of class names for variants
This should make the code slightly faster as Polymer does not need to do
string interpolation for CSS classes. In practice, this does not
noticeably affect emoji picker load speed.

This also cleans up the various `hasVariants` code, consolidating the
two implementations of checking whether an `EmojiVariants` has variants,
and removes the redundant "has variants" check in the template when we
know that a group has no variants.

A similar change was suggested in
https://crrev.com/c/3901656/comment/7ddac14e_8347b30c/.

Bug: None
Change-Id: I3bcadf4032b9049590114ae50e001762a8ad804b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5239660
Reviewed-by: John Palmer <jopalmer@chromium.org>
Commit-Queue: Michael Cui <mlcui@google.com>
Cr-Commit-Position: refs/heads/main@{#1254338}