Remove no longer needed raw_ptr rewrite path exclusions
These files were excluded, because raw_ptr and PA depended on them, thus creating a cyclic dependency. Both are now in a standalone library that have their own copy of `base`.

Instead, we need to add manual exclusions in lock_impl.h

This CL removes the exclusions from the clang plugin that enforces
raw_ptr usage, but adds them to the Chromium-side plugin params. This
is to avoid breakage in case new T* show up in these files by the time
the plugin is rolled in.

Change-Id: I70199115e5904fbf4e21798f02ad627f1aea9b01
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5205312
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Commit-Queue: Bartek Nowierski <bartekn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254379}