[accessibility] Measure the time to send mode flags to renderers
Record the round-trip time to send a set of accessibility mode flags
from the browser process to a renderer and receive and process the first
set of updates and events. Accessibility.EventProcessingTime.First holds
this round-trip time for the first call when the binding to
RenderAccessibility takes place. Accessibility.EventProcessingTime holds
the time for all other calls.

Bug: 1470199
Change-Id: I51dd109996e928799bfece6b3bf7897cf61c28c3
AX-Relnotes: n/a.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5244565
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Commit-Queue: Greg Thompson <grt@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254419}