[CrashFix] Fix password check usage in ChromeNativePasswordCheckController
There were multiple crashes reported when starting the password check.
Assumption: ChromeNativePasswordCheckController is a long living object
in the SafetyCheckMediator. It might be, that somehow the PasswordCheck
is re-created during the controller's lifetime and this leads to calling
the password check on the object whose C++ counterpart was already
destroyed.
Fix: Similar to how it's done in the PasswordCheckMediator (link
https://source.chromium.org/chromium/chromium/src/+/main:chrome/browser/password_check/android/internal/java/src/org/chromium/chrome/browser/password_check/PasswordCheckMediator.java;l=349;drc=b58207f5aecc39db7d3da766e7d171e5d2c3598e)
do not store the password check object. Instead, query it every time
from the password check factory when it is needed.

Bug: 1523476
Change-Id: I30f83bda88e1db51ee9d64e08d2e5bd600882bc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5253367
Commit-Queue: Anna Tsvirchkova <atsvirchkova@google.com>
Reviewed-by: Friedrich Horschig <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254444}