[iOSKAUpgrade] Fix header alignment in PasswordViewController
crrev.com/c/5237988, introduced an issue where, when the Keyboard
Accessory Upgrade feature is disabled, the PasswordViewController's
header is taking the whole width of the table view, and is hence not
aligned with the rest of the content.

This CL fixes the issue by adding a leading and a trailing indent to the
PasswordViewController's header when the Keyboard Accessory Upgrade
feature is disabled.

Bug: 1501851
Change-Id: I798e0e2bb5578a94271e87199599614b64c4221c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5248994
Commit-Queue: Noémie St-Onge <noemies@google.com>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254515}