[UNO] Record `ShouldShowChromeSigninBubbleWithReason` more
Currently
`Signin.Intercept.Heuristic.ShouldShowChromeSigninBubbleWithReason` will not be recorded if:
(1) the information required to compute `GetHeuristicOutcome()`
is not fully available (e.g. `EnterpriseSeparationMaybeRequired()` return
nullopt).
(2) and `ShouldShowChromeSigninBubble()` returns false.

This gives less data on the reasons why chrome sign in bubble was not
shown. It is also more correct to record the decision made in
`OnInterceptionReadyToBeProcessed()`.

This CL ensures the metric is recorded if early abort or in
`OnInterceptionReadyToBeProcessed()` on success and failure.

Bug: b/b/320426816
Change-Id: I94c9c28f8fb52e55f8e99139a3abcbea251a1cd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5245606
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Monica Basta <msalama@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254516}