[infra] Traverse edges to compute compile targets for a bundle.
Currently the compile targets for a bundle are stored as a property on
the bundle in addition to adding edges to their nodes. This
unnecessarily duplicates information. Since the edges provide the
guarantee that the compile targets have been declared, this changes it
to get the compile targets by traversing the edges and removes the extra
property.

Bug: 1420012
Change-Id: I2520c98fbe8acf0d33d4881339b1d39bd5966a81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5250188
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Cr-Commit-Position: refs/heads/main@{#1254570}