[a11y] SetValue on a video's shadow tree <input type=range> now works
We had found a bug where kSetValue action when performed on the
<input type=range> element that is the timeline scrubber for a video
element was not working.

This was because this element internally was a AXMediaControl,
which unlike AXSlider does not have an implementation for
OnNativeSetValueAction and so we were defaulting to
AXNodeObject::OnNativeSetValueAction which led then to incorrect
behavior since we would expect this element to behave like a AXSlider
in this case.

This CL fixes this by adding an implementation for
OnNativeSetValueAction on AXMediaControl so that it behaves exactly
like AXSlider::OnNativeSetValueAction when appropriate.

Change-Id: Ia802f7ab8fe13e5a506b973f864a9dc6ff3f2fca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5124457
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1254588}