Fix check action for fuzztests on linux32.
We have an action which checks that the fuzztest = [...] gn line is
correct. This requires running the fuzzer binary, and sometimes we can't
do that - for example if a certain dependency DSO is missing in the CI
environment. In such cases we assume that the 'fuzztests' line is fine.
But in this circumstance we weren't writing to disk the 'checked.txt'
file which communicated to ninja that the action had been performed, and
thus ninja thought it needed to run this again on the next build.

Bug: 1499204
Cq-Include-Trybots: luci.chromium.try:linux-centipede-asan-rel
Cq-Include-Trybots: luci.chromium.try:linux-libfuzzer-asan-rel
Cq-Include-Trybots: luci.chromium.try:win-libfuzzer-asan-rel
Change-Id: I2e9a72284facbbc0d052fb2f985fff9ae5ce54da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5253275
Reviewed-by: Paul Semel <paulsemel@chromium.org>
Commit-Queue: Adrian Taylor <adetaylor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254603}