Browser ignores intents on multi-display devices if main display is off
Commit c5b50f4c2d42e introduced a test that causes the browser to ignore
intents if the display is off. The test is implemented in
`chrome/android/java/src/org/chromium/chrome/browser/IntentHandler.java`
by calling the `isInteractive()` method of `PowerManageer`.

The approach works when there is only one display. However, it makes
the browser unusable in devices capable of turning on and off displays
independently of each other. An example are Samsung devices running
Samsung DeX. If the user is running the browser on the external display
and the main display is off, the browser will ignore all intents. This
prevents apps running on the external display, such as a mail client,
from opening webpages when the user clicks on a link.

The solution is to modify the `isScreenOn()` method to use
`Display#getState()` on the display returned by
`WindowManager#getDefaultDisplay()`.

This issue affects both Chromium and Chrome.

Bug: 1522543

display, open the browser on the external monitor, verify that tabs
can be opened by sending intents; see crbug.com/1522543 for more details.

Test: Start Samsung DeX on an external monitor, turn off the main
Change-Id: I8a48b546bbfc1f19d4738faf9acd9aeb0251567a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5245736
Commit-Queue: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@google.com>
Cr-Commit-Position: refs/heads/main@{#1254617}