[dPWA] Fix flakiness in WebAppSyncBridge unit-tests
This CL fixes the flakiness in the WebAppSyncBridge unit-test for
deleting sync apps, where deletion was "mocked" by calling DeleteApp()
on the sync bridge layer directly, leading to UAFs elsewhere in the
web_applications/ code due to the absence of locks on the web app
being deleted. By wrapping the deletion in a command, the problem is solved.

Bug: 1521168
Change-Id: I91a804a90b57073738e30ec7a27c4eea38eb7037
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5249632
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Auto-Submit: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254632}