[RQ] Always filter out SRP urls from the MV sites
Currently max 1 SRP url can show up in the MV sites if the SRP segment
(e.g., http://google.com/search) makes it way into the segments table
(e.g., via bookmark click). This causes the latest search to always show
up in the MV sites if the accumulated visit count of that segment is
high enough. This is not a good experience and has been causing a lot of
confusion.

SRP URLs are filtered out from the MV sites when the Repeatable Queries
are enabled which is the case on Android only. This CL always filters
out those sites, whether or not Repeatable Queries are enabled.

Bug: 1317829
Change-Id: I3a2075e8a1f4150b5f1e7938da1c5b9573719e9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5249631
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254702}