[Nearby Presence] Create PresenceDevice map to keep pointers in scope
Because |Core| takes in a `const PresenceDevice&`, it does not maintain
its own copy of PresenceDevice. Once the PresenceDevice goes out of
scope, referencing it in Core will cause a segfault. This change instead
uses a `PresenceDevice*` flat_map so Core won't be holding invalid
references when these functions go out of scope.

Test: this is an implementation detail, so no unit tests were added as
this is not part of the NC interface.

Fixed: b/315016880
Change-Id: I380ee6961468181f07b395d5a9a37f54fe1b4449
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5113403
Reviewed-by: Ryan Hansberry <hansberry@chromium.org>
Commit-Queue: Crisrael Lucero <crisrael@google.com>
Reviewed-by: Juliet Lévesque <julietlevesque@google.com>
Cr-Commit-Position: refs/heads/main@{#1254737}