views: don't clip modal dialog to its parent
The modal dialog was clipped to its host's boundary. This caused problem
in mac immersive fullscreen because the host is dedicated to the tab
strip and as such the dialog is clipped to the tab strip area.

This CL keeps modal dialog's original size. The clipping logic was
adding in crbug.com/1341530. I manually tested the bug and can confirm
that this change does not resurrect that bug.

Bug: 1521451
Change-Id: Ie42c13a034374123436c7c07f841d5c35fd62b68
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5253739
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Commit-Queue: Keren Zhu <kerenzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254747}