Remove blink::MainThreadIsolate
For the MBI project we need to be able to handle multiple isolates
on the main thread. Most code was rewritten to handle this. The last
usage is TaskEnvironment, and we can look at the MainThreadScheduler
in that case because it is progressively being rewritten to use the
variant that creates a new isolate each TaskEnvironment.

Bug: 263412
Change-Id: Ic608539395e036197f22f679c17f2f570cf39b77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5249640
Commit-Queue: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254764}